Rename package `scs_codec` to `scs`.
This change simplifies the package name to better align with its purpose and usage. Corresponding updates have been made to the source code, tests, and documentation to reflect this change.
